fix: images not fitting a layout aren't dropped
Previously when there was no layout that would fit the remaining images,
those would have been dropped. Now the OneFourLayout is used as fallback
for such cases. Based on the way the layout is created and the fact that
the images end up roughly square in the final layout, this allows for
a decent layout when a low number of images is left.