feat: new layout for vertical images
The main reason behind creating this layout is that a batch of 4
vertical images did not have a fitting layout yet. This leads to an
infinite loop attempting to find a fitting layout for this case which
fails to complete.
Adding this layout that takes at least 2 vertical images and has no
required horizontal images this problem is solved.