chore: refactored chosing a layout
Previously after a layout was chosen and the images didn't fit the
requirements of the layout regarding orientation of the images, the
images were put back into the list of images and an empty component was
returned to update the PhotoWall which in turn gave the random generator
another shot at trying to find a fitting layout for the images.
This was changed to actually search for a fitting layout now to prevent
multiple iterations of this loop returning empty components just because
the images didn't fit.