fix: free spots in photowall
There was an issue that could cause free spots in the photowall to occur
in rare cases. This can happen if the picked images don't fit the
layout's requirements and a layout with more free spots is chosen.

Layouts are now additionally filtered for total images required to match
the amount of picked images to fix this.